+++ src/history/index.ts
@@ -1,1 +1,2 @@
+export * from "./CacheChange";
 export * from "./HistoryCache";
