+++ src/index.ts
@@ -4,7 +4,6 @@ export * from "./messaging";
 export * from "./routing";
 export * from "./transport";
 
-export * from "./Endpoint";
 export * from "./Participant";
 export * from "./ParticipantAttributes";
 export * from "./ParticipantView";
